Feat: Added message queue and its thread and functions

See merge request yonhag/netanya-1102-tor!8

+++ Proxy/Relay.h
@@ -3,7 +3,7 @@
 
 struct Relay
 {
-	unsigned int _publicRSAKey;
-	unsigned int _publicAESKey;
+	unsigned long _publicRSAKey;
+	unsigned long _publicAESKey	;
 	std::string _ip;
 };
